life-coach-agent/pyproject.toml
[project]
name = "life-coach-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.13.6",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
